move useEffect from context into hook
