+++ src/hooks/useMap.js
@@ -3,6 +3,7 @@ import { isEqual } from 'lodash';
 
 import { MapContext } from 'contexts/MapContext';
 import TopicsService from 'services/TopicsService';
+import TargetsService from 'services/TargetsService';
 
 const useMap = () => {
   const {
@@ -65,6 +66,33 @@ const useMap = () => {
 
     fetchTopics();
   }, [topicsList]);
+
+  useEffect(() => {
+    const fetchCurrentLocation = () => {
+      navigator.geolocation.getCurrentPosition((position) => {
+        if (mapProperties.location) return;
+
+        setMapProperties({
+          ...mapProperties,
+          location: {
+            lat: position.coords.latitude,
+            lng: position.coords.longitude,
+          },
+        });
+      });
+    };
+
+    const fetchTargets = async () => {
+      const response = await TargetsService.getTargets();
+      const targetsIds = targets.map(({ id }) => id).sort();
+      const responseIds = response.map(({ id }) => id).sort();
+
+      if (!isEqual(targetsIds, responseIds)) setTargets(response);
+    };
+
+    fetchCurrentLocation();
+    fetchTargets();
+  }, [mapProperties, targets, setMapProperties, setTargets]);
 
   const handleMapClick = (e) => {
     setMapProperties({
